Pull out Activation steps into ActivationStack
